+++ contracts/runtime-upgrade/src/lib.rs
@@ -3,13 +3,14 @@
 
 extern crate alloc;
 
+#[cfg(test)]
+mod tests;
+
 use alloc::{borrow::Cow, string::String, vec, vec::Vec};
 use fluentbase_sdk::{
-    basic_entrypoint,
-    codec::Codec,
-    compile_rwasm_maybe_system,
+    basic_entrypoint, compile_rwasm_maybe_system,
     crypto::crypto_keccak256,
-    derive::{function_id, router, Contract, Event},
+    derive::{router, Contract, Event},
     hex,
     storage::{StorageAddress, StorageBytes32, StorageString, StorageVec},
     syscall::{encode, SYSCALL_ID_UPGRADE_RUNTIME},
@@ -41,7 +42,7 @@ struct UpgradePlanned {
     genesis_version: String,
     target_addresses: Vec<Address>,
     wasm_code_hashes: Vec<B256>,
-    upgrador: Address,
+    updater: Address,
 }
 
 #[derive(Event)]
@@ -55,7 +56,7 @@ struct App<SDK> {
     owner: StorageAddress,
     planned_genesis_hash: StorageBytes32,
     planned_genesis_version: StorageString,
-    planned_upgrador: StorageAddress,
+    planned_updater: StorageAddress,
     planned_target_addresses: StorageVec<StorageAddress>,
     planned_wasm_hashes: StorageVec<StorageBytes32>,
 }
@@ -83,7 +84,7 @@ trait RuntimeUpgradeTr {
         genesis_version: String,
         target_addresses: Vec<Address>,
         wasm_code_hashes: Vec<B256>,
-        upgrador: Address,
+        updater: Address,
     );
 
     /// Upgrade WASM runtime smart contract using a previously planned target/hash pair.
@@ -159,7 +160,7 @@ impl<SDK: SharedAPI> RuntimeUpgradeTr for App<SDK> {
         genesis_version: String,
         target_addresses: Vec<Address>,
         wasm_code_hashes: Vec<B256>,
-        upgrador: Address,
+        updater: Address,
     ) {
         _ = self.only_owner();
         if wasm_code_hashes.is_empty() {
@@ -168,12 +169,12 @@ impl<SDK: SharedAPI> RuntimeUpgradeTr for App<SDK> {
         if target_addresses.len() != wasm_code_hashes.len() {
             panic!("runtime-upgrade: mismatched upgrade plan");
         }
-        if upgrador == Address::ZERO {
-            panic!("runtime-upgrade: planned upgrador is zero address");
+        if updater == Address::ZERO {
+            panic!("runtime-upgrade: planned updater is zero address");
         }
 
         // Validate the whole replacement plan before clearing the previous one. The runtime should
-        // never persist a partial plan if one entry is malformed.
+        // never persist as a partial plan if one entry is malformed.
         for (index, (target_address, wasm_code_hash)) in target_addresses
             .iter()
             .copied()
@@ -196,8 +197,7 @@ impl<SDK: SharedAPI> RuntimeUpgradeTr for App<SDK> {
             .set(&mut self.sdk, genesis_hash);
         self.planned_genesis_version_accessor()
             .set(&mut self.sdk, &genesis_version);
-        self.planned_upgrador_accessor()
-            .set(&mut self.sdk, upgrador);
+        self.planned_updater_accessor().set(&mut self.sdk, updater);
 
         for (target_address, wasm_code_hash) in target_addresses
             .iter()
@@ -215,7 +215,7 @@ impl<SDK: SharedAPI> RuntimeUpgradeTr for App<SDK> {
             genesis_version,
             target_addresses,
             wasm_code_hashes,
-            upgrador,
+            updater,
         }
         .emit(&mut self.sdk)
         .unwrap();
@@ -223,13 +223,13 @@ impl<SDK: SharedAPI> RuntimeUpgradeTr for App<SDK> {
 
     #[function_id("upgradeToPlanned(address,bytes)")]
     fn upgrade_to_planned(&mut self, target_address: Address, wasm_bytecode: Bytes) {
-        let upgrador = self.planned_upgrador_accessor().get(&self.sdk);
-        if upgrador == Address::ZERO {
+        let updater = self.planned_updater_accessor().get(&self.sdk);
+        if updater == Address::ZERO {
             panic!("runtime-upgrade: no planned upgrade");
         }
         let caller = self.sdk.context().contract_caller();
-        if caller != upgrador {
-            panic!("runtime-upgrade: incorrect planned upgrador");
+        if caller != updater {
+            panic!("runtime-upgrade: incorrect planned updater");
         }
 
         let wasm_code_hash = crypto_keccak256(wasm_bytecode.as_ref());
@@ -381,168 +381,3 @@ impl<SDK: SharedAPI> App<SDK> {
 }
 
 basic_entrypoint!(App);
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use fluentbase_sdk::{address, bytes, ContractContextV1, ExitCode, B256};
-    use fluentbase_testing::TestingContextImpl;
-
-    struct Harness {
-        sdk: TestingContextImpl,
-    }
-
-    impl Harness {
-        fn new() -> Self {
-            Self {
-                sdk: TestingContextImpl::default().with_contract_context(ContractContextV1 {
-                    gas_limit: 120_000,
-                    ..Default::default()
-                }),
-            }
-        }
-
-        fn set_caller(&mut self, caller: Address) {
-            self.sdk.context_mut().caller = caller;
-        }
-
-        fn call<I: Into<Bytes>>(&mut self, input: I) -> ExitCode {
-            self.sdk = core::mem::take(&mut self.sdk).with_input(input.into());
-            let storage_before_call = self.sdk.dump_storage();
-            let mut app = App::new(core::mem::take(&mut self.sdk));
-            let exit_code =
-                match std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| app.main())) {
-                    Ok(_) => ExitCode::Ok,
-                    Err(_) => ExitCode::Panic,
-                };
-            self.sdk = app.sdk;
-            if !exit_code.is_ok() {
-                self.sdk.restore_storage(storage_before_call);
-            }
-            _ = self.sdk.take_output();
-            exit_code
-        }
-    }
-
-    #[test]
-    fn test_upgrade_to_encoding() {
-        let target = address!("2222222222222222222222222222222222222222");
-        let genesis_hash = B256::from([0xab; 32]);
-        let genesis_version = "v1.0.0".to_string();
-        // minimal valid WASM: magic bytes + version
-        let wasm_bytecode = Bytes::from([0x00, 0x61, 0x73, 0x6d, 0x01, 0x00, 0x00, 0x00].as_ref());
-
-        let call = UpgradeToCall::new((
-            target,
-            genesis_hash,
-            genesis_version.clone(),
-            wasm_bytecode.clone(),
-        ));
-        let encoded = call.encode();
-
-        // first 4 bytes = function selector
-        assert!(encoded.len() >= 4);
-        println!("Encoded call data: {}", hex::encode(&encoded));
-
-        // decode back and verify a round-trip
-        let decoded = UpgradeToCall::decode(&&encoded[4..]).expect("failed to decode");
-        assert_eq!(decoded.0 .0, target, "target_address mismatch");
-        assert_eq!(decoded.0 .1, genesis_hash, "genesis_hash mismatch");
-        assert_eq!(decoded.0 .2, genesis_version, "genesis_version mismatch");
-        assert_eq!(decoded.0 .3, wasm_bytecode, "wasm_bytecode mismatch");
-    }
-
-    #[test]
-    fn test_recompile_encoding() {
-        let target = address!("2222222222222222222222222222222222222222");
-
-        let call = RecompileCall::new((target,));
-        let encoded = call.encode();
-
-        assert!(encoded.len() >= 4);
-        println!("Encoded call data: {}", hex::encode(&encoded));
-
-        let decoded = RecompileCall::decode(&&encoded[4..]).expect("failed to decode");
-        assert_eq!(decoded.0 .0, target, "target_address mismatch");
-    }
-
-    #[test]
-    fn test_plan_upgrade_encoding() {
-        let genesis_hash = B256::from([0xab; 32]);
-        let genesis_version = "v1.0.0".to_string();
-        let target_addresses = vec![
-            address!("2222222222222222222222222222222222222222"),
-            address!("3333333333333333333333333333333333333333"),
-        ];
-        let wasm_code_hashes = vec![B256::from([0x11; 32]), B256::from([0x22; 32])];
-        let upgrador = address!("1111111111111111111111111111111111111111");
-
-        let call = PlanUpgradeCall::new((
-            genesis_hash,
-            genesis_version.clone(),
-            target_addresses.clone(),
-            wasm_code_hashes.clone(),
-            upgrador,
-        ));
-        let encoded = call.encode();
-
-        assert!(encoded.len() >= 4);
-        let decoded = PlanUpgradeCall::decode(&&encoded[4..]).expect("failed to decode");
-        assert_eq!(decoded.0 .0, genesis_hash, "genesis_hash mismatch");
-        assert_eq!(decoded.0 .1, genesis_version, "genesis_version mismatch");
-        assert_eq!(decoded.0 .2, target_addresses, "target_addresses mismatch");
-        assert_eq!(decoded.0 .3, wasm_code_hashes, "wasm_code_hashes mismatch");
-        assert_eq!(decoded.0 .4, upgrador, "upgrador mismatch");
-    }
-
-    #[test]
-    fn test_upgrade_to_planned_encoding() {
-        let target = address!("2222222222222222222222222222222222222222");
-        let wasm_bytecode = Bytes::from([0x00, 0x61, 0x73, 0x6d, 0x01, 0x00, 0x00, 0x00].as_ref());
-
-        let call = UpgradeToPlannedCall::new((target, wasm_bytecode.clone()));
-        let encoded = call.encode();
-
-        assert!(encoded.len() >= 4);
-        let decoded = UpgradeToPlannedCall::decode(&&encoded[4..]).expect("failed to decode");
-        assert_eq!(decoded.0 .0, target, "target_address mismatch");
-        assert_eq!(decoded.0 .1, wasm_bytecode, "wasm_bytecode mismatch");
-    }
-
-    #[test]
-    fn test_upgrade_and_recompile_event_signatures_are_distinct() {
-        assert_eq!(
-            RuntimeUpgraded::SIGNATURE,
-            "RuntimeUpgraded(address,bytes32,string,bytes32)"
-        );
-        assert_eq!(
-            ContractRecompiled::SIGNATURE,
-            "ContractRecompiled(address,bytes32)"
-        );
-        assert_ne!(RuntimeUpgraded::SELECTOR, ContractRecompiled::SELECTOR);
-    }
-
-    #[test]
-    fn test_planned_upgrade_rejects_same_hash_for_wrong_target() {
-        let planned_target = address!("2222222222222222222222222222222222222222");
-        let wrong_target = address!("3333333333333333333333333333333333333333");
-        let upgrador = address!("1111111111111111111111111111111111111111");
-        let wasm_bytecode = Bytes::from([0x00, 0x61, 0x73, 0x6d, 0x01, 0x00, 0x00, 0x00].as_ref());
-        let wasm_code_hash = crypto_keccak256(wasm_bytecode.as_ref());
-
-        let mut h = Harness::new();
-        h.set_caller(DEFAULT_UPDATE_GENESIS_AUTH);
-        let plan_call = PlanUpgradeCall::new((
-            B256::from([0xab; 32]),
-            "v1.0.0".to_string(),
-            vec![planned_target],
-            vec![wasm_code_hash],
-            upgrador,
-        ));
-        assert_eq!(h.call(plan_call.encode()), ExitCode::Ok);
-
-        h.set_caller(upgrador);
-        let upgrade_call = UpgradeToPlannedCall::new((wrong_target, wasm_bytecode));
-        assert_ne!(h.call(upgrade_call.encode()), ExitCode::Ok);
-    }
-}
